+++ src/VulkanRenderer/Shader/ShaderManager.cpp
@@ -18,7 +18,7 @@ std::vector<char> ShaderManager::getBinaryDataFromFile(const std::string& filena
     if (!file.is_open())
         throw std::runtime_error("Failed to open file: " + filename + ".spv");
 
-    size_t fileSize = (size_t)file.tellg();
+    uint32_t fileSize = (uint32_t)file.tellg();
     std::vector<char> buffer(fileSize);
 
     // Seeks back to the beginning of the file.
